Add react rendered block that handles overlays
By publishing the new event that was added in CMS.UI.Core 11.2.0 we can
tell the editing ui when they need to update their overlays so that on
page editing (OPE) works when react has re-rendered a property.

Adding the ThreeKeyFacts component that demonstrates this by showing
different properties when the user clicks buttons and publishing the
event to make sure that the OPE editing works for the property currently
being shown.

PreviewController get a [RequireClientResources] attribute since that is
required to get access to the global epi object in the templates to
publish the new event when we are editing blocks.